+++ src/SatisfiabilityInterface.jl
@@ -18,15 +18,15 @@ import CommonSolve: solve
 
 include("symbolics_interface.jl")
 include("discrete_variables.jl")
-
+include("nodes.jl")
 include("sat_problem.jl")
 include("symbolic_problem.jl")
 include("MOI_wrapper.jl")
 include("read_cnf.jl")
 include("solver.jl")
-# include("encode.jl")
+include("parsing.jl")
 
 include("relations.jl")
-
+include("CSP.jl")
 
 end
